requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
dataclasses-json==0.5.8
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
keyboard==0.13.5
langchain==0.0.311
langchainplus-sdk==0.0.17
langsmith==0.0.43
marshmallow==3.19.0
marshmallow-enum==1.5.1
mistune==3.0.1
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.25.0
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
pydantic==1.10.9
pyobjc==9.2
pyobjc-core==9.2
pyobjc-framework-Accessibility==9.2
pyobjc-framework-Accounts==9.2
pyobjc-framework-AddressBook==9.2
pyobjc-framework-AdServices==9.2
pyobjc-framework-AdSupport==9.2
pyobjc-framework-AppleScriptKit==9.2
pyobjc-framework-AppleScriptObjC==9.2
pyobjc-framework-ApplicationServices==9.2
pyobjc-framework-AppTrackingTransparency==9.2
pyobjc-framework-AudioVideoBridging==9.2
pyobjc-framework-AuthenticationServices==9.2
pyobjc-framework-AutomaticAssessmentConfiguration==9.2
pyobjc-framework-Automator==9.2
pyobjc-framework-AVFoundation==9.2
pyobjc-framework-AVKit==9.2
pyobjc-framework-BusinessChat==9.2
pyobjc-framework-CalendarStore==9.2
pyobjc-framework-CallKit==9.2
pyobjc-framework-CFNetwork==9.2
pyobjc-framework-ClassKit==9.2
pyobjc-framework-CloudKit==9.2
pyobjc-framework-Cocoa==9.2
pyobjc-framework-Collaboration==9.2
pyobjc-framework-ColorSync==9.2
pyobjc-framework-Contacts==9.2
pyobjc-framework-ContactsUI==9.2
pyobjc-framework-CoreAudio==9.2
pyobjc-framework-CoreAudioKit==9.2
pyobjc-framework-CoreBluetooth==9.2
pyobjc-framework-CoreData==9.2
pyobjc-framework-CoreHaptics==9.2
pyobjc-framework-CoreLocation==9.2
pyobjc-framework-CoreMedia==9.2
pyobjc-framework-CoreMediaIO==9.2
pyobjc-framework-CoreMIDI==9.2
pyobjc-framework-CoreML==9.2
pyobjc-framework-CoreMotion==9.2
pyobjc-framework-CoreServices==9.2
pyobjc-framework-CoreSpotlight==9.2
pyobjc-framework-CoreText==9.2
pyobjc-framework-CoreWLAN==9.2
pyobjc-framework-CryptoTokenKit==9.2
pyobjc-framework-DeviceCheck==9.2
pyobjc-framework-DictionaryServices==9.2
pyobjc-framework-DiscRecording==9.2
pyobjc-framework-DiscRecordingUI==9.2
pyobjc-framework-DiskArbitration==9.2
pyobjc-framework-DVDPlayback==9.2
pyobjc-framework-EventKit==9.2
pyobjc-framework-ExceptionHandling==9.2
pyobjc-framework-ExecutionPolicy==9.2
pyobjc-framework-ExternalAccessory==9.2
pyobjc-framework-FileProvider==9.2
pyobjc-framework-FileProviderUI==9.2
pyobjc-framework-FinderSync==9.2
pyobjc-framework-FSEvents==9.2
pyobjc-framework-GameCenter==9.2
pyobjc-framework-GameController==9.2
pyobjc-framework-GameKit==9.2
pyobjc-framework-GameplayKit==9.2
pyobjc-framework-ImageCaptureCore==9.2
pyobjc-framework-IMServicePlugIn==9.2
pyobjc-framework-InputMethodKit==9.2
pyobjc-framework-InstallerPlugins==9.2
pyobjc-framework-InstantMessage==9.2
pyobjc-framework-Intents==9.2
pyobjc-framework-IOBluetooth==9.2
pyobjc-framework-IOBluetoothUI==9.2
pyobjc-framework-IOSurface==9.2
pyobjc-framework-iTunesLibrary==9.2
pyobjc-framework-KernelManagement==9.2
pyobjc-framework-LatentSemanticMapping==9.2
pyobjc-framework-LaunchServices==9.2
pyobjc-framework-libdispatch==9.2
pyobjc-framework-libxpc==9.2
pyobjc-framework-LinkPresentation==9.2
pyobjc-framework-LocalAuthentication==9.2
pyobjc-framework-MapKit==9.2
pyobjc-framework-MediaAccessibility==9.2
pyobjc-framework-MediaLibrary==9.2
pyobjc-framework-MediaPlayer==9.2
pyobjc-framework-MediaToolbox==9.2
pyobjc-framework-Metal==9.2
pyobjc-framework-MetalKit==9.2
pyobjc-framework-MetalPerformanceShaders==9.2
pyobjc-framework-MetalPerformanceShadersGraph==9.2
pyobjc-framework-MLCompute==9.2
pyobjc-framework-ModelIO==9.2
pyobjc-framework-MultipeerConnectivity==9.2
pyobjc-framework-NaturalLanguage==9.2
pyobjc-framework-NetFS==9.2
pyobjc-framework-Network==9.2
pyobjc-framework-NetworkExtension==9.2
pyobjc-framework-NotificationCenter==9.2
pyobjc-framework-OpenDirectory==9.2
pyobjc-framework-OSAKit==9.2
pyobjc-framework-OSLog==9.2
pyobjc-framework-PassKit==9.2
pyobjc-framework-PencilKit==9.2
pyobjc-framework-Photos==9.2
pyobjc-framework-PhotosUI==9.2
pyobjc-framework-PreferencePanes==9.2
pyobjc-framework-PushKit==9.2
pyobjc-framework-Quartz==9.2
pyobjc-framework-QuickLookThumbnailing==9.2
pyobjc-framework-ReplayKit==9.2
pyobjc-framework-SafariServices==9.2
pyobjc-framework-SceneKit==9.2
pyobjc-framework-ScreenSaver==9.2
pyobjc-framework-ScreenTime==9.2
pyobjc-framework-ScriptingBridge==9.2
pyobjc-framework-SearchKit==9.2
pyobjc-framework-Security==9.2
pyobjc-framework-SecurityFoundation==9.2
pyobjc-framework-SecurityInterface==9.2
pyobjc-framework-ServiceManagement==9.2
pyobjc-framework-Social==9.2
pyobjc-framework-SoundAnalysis==9.2
pyobjc-framework-Speech==9.2
pyobjc-framework-SpriteKit==9.2
pyobjc-framework-StoreKit==9.2
pyobjc-framework-SyncServices==9.2
pyobjc-framework-SystemConfiguration==9.2
pyobjc-framework-SystemExtensions==9.2
pyobjc-framework-UniformTypeIdentifiers==9.2
pyobjc-framework-UserNotifications==9.2
pyobjc-framework-UserNotificationsUI==9.2
pyobjc-framework-VideoSubscriberAccount==9.2
pyobjc-framework-VideoToolbox==9.2
pyobjc-framework-Virtualization==9.2
pyobjc-framework-Vision==9.2
pyobjc-framework-WebKit==9.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.30.0
sniffio==1.3.0
SQLAlchemy==2.0.17
SQLite3-0611==0.0.1
tenacity==8.2.2
tqdm==4.65.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.0
urllib3==2.0.2
yarl==1.9.2
